requirements.txt
flask
opencv-python
mediapipe
numpy
pandas
scikit-learn
flask_sqlalchemy
flask_migrate
flask_login
flask_wtf
wtforms
pymysql
werkzeug
openai-whisper
torch
torchaudio
